+++ resources/views/user/list.blade.php
@@ -75,7 +75,9 @@
 @include('user.dialog.add-tags-dialog')
 @include('user.dialog.description-dialog')
 @include('report.dialog.user-report-action')
-@include('layouts.paginate', ['collection' => $users, 'per_page' => isset($per_page)? $per_page : ''])
+<div class="text-center">
+    {!! $users->appends(Input::all())->render() !!}
+</div>
 
 @stop